// src/pages/quizzes/types/theme.type.ts
export interface IQuizTheme {
  primaryTextClass: string;
  secondaryTextClass: string;
  secondaryActiveTextClass: string;
  numberTextClass: string;
  numberBackgroundClass: string;
  backgroundClass: string;
  btnsBackgroundClass: string;
  btnsTextClass: string;
  radioButtonColor: string;
  titleTextClass: string;
  personName: string;
}